intellicloud-backend-backup/requirements.backup.txt
Flask==3.1.1
flask-cors==6.0.0
firebase-admin==6.8.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
requests==2.32.3
PyJWT==2.10.1
